requirements.txt
python-dotenv==0.20.0
schedule==1.1.0
secret==0.8
tweepy==4.10.0
